Jump to callee when gotodef on keyword argument
Summary:
This diff starts to add handling logic when gotodef on a keyword arg.

To get things started, we would simply jump to the corresponding callee for now. This is made easier with D76375057 and D76375056 which provide convenient APIs for us to issue gotodef requests to a given identifier or an attribute access.

In the next diff we are going to further "refine" the location by parameter name.

Reviewed By: SamChou19815

Differential Revision: D76379066

fbshipit-source-id: 94a34bd07e2d23eb37b606243308ae0d7ea8521c