Add parameter context to `identifier_at()`
Summary:
shamecube

It seems like I fogot to code up an important case before removing the call to `definition_at()` in D76358458 -- when the user cmd+click on the exact name of a parameter.

This diff attempts to adds the case back and let those cases jump to themselves. I think for gotodef purpose, we don't need to differentiate what kind of parameters we are looking at since that doesn't matter. For `get_type_at()` though the story might be more complicated, but I'm focusing on gotodef for now.

Reviewed By: SamChou19815

Differential Revision: D76393323

fbshipit-source-id: 1c023eb1f14f3de220eef5d8cd5f8d4fb1a4f3db