Add Key/Binding plumbing to pin upstream placeholders
Summary:
This is the first half of closing the leakage of placeholder types through first
use chains - create the Key/Binding variants we need.

We'll use these to separate the raw binding (which contains unpinned vars that
might originate from ourselves or upstream) versus the binding we'll expose to
our first uses (which contains only the ones originating from ourselves).

Reviewed By: yangdanny97

Differential Revision: D76273079

fbshipit-source-id: f91518c0fb9c10977543428a2833cc363bd86fd9