Pin placeholders everywhere *except* NameAssign
Summary:
We want placeholder types in Pyrefly to be used *exclusively* for
name assignments, where they will get usage tracked. All other bindings
should immediately force any (non-recursive) vars they encounter.

This nearly completes my current effort on placeholder type pinning; the
only remaining source of unpinned vars potentially leaking into
nondeterministic export-forced calculations that I know of is the case
where a first-usage is also a `NameAssign`, which would lead to *its* first
usage being able to see `Var`s introduce "two first usages" ago.

The only way I know of to fix that is to introduce an extra layer of indirection
when we are in a `NameAssign` that is also a first usage, which I'll need
to work on next.

There may also be some unknown var leaks, in particular if we are duplicating
expressions of first usages into multiple bindings (as we were for narrowing until
I recently marked narrows as "non-pinning" uses).

Reviewed By: yangdanny97

Differential Revision: D76273086

fbshipit-source-id: 463401fe8290c8926889cc578c0118b3935e7006