Pass `Usage` as `&mut`
Summary:
This is another step toward being able to track what dependencies
might have had a `Pin` impacted on `NameAssign` bindings, which is needed to
prevent leaking placeholder vars and introducing nondeterminism for usage chains.

In order to track these dependencies, the obvious approach will be to attach
them to the usage which is easiest to do if we pass it by mutable ref. Good news is,
as this diff shows that does not cause us any borrow tracker issues :)

Reviewed By: yangdanny97

Differential Revision: D76273081

fbshipit-source-id: 7d130e1caf42b224fb441256702d05c3993602f6