Remove some traits from Usage
Summary:
At the time I wrote `Usage`, I was thinking I might store it directly
in a `Binding` (my thinking was that having a bunch of `Idx<Key>` that
don't tell you what they mean, and some comments, is not as good as
having well-named types).

I didn't wind up doing that though, and now I need Usage to start
tracking reverse dependencies which doesn't play well with some of these
traits that we aren't using. Time to get rid of the unnecessary traits.

Reviewed By: yangdanny97

Differential Revision: D76273087

fbshipit-source-id: c199f449d44dbf8f530ba498cc9e9040220e0fd2