Add a `Pin` binding
Summary:
For first-usage based pinning of placeholder types within the existing
graph structure of `Bindings`, we need a way to "trap" placeholder types in
a binding that can only be seen by a first usage. All other usages will flow
through a binding that *depends* on the first usage, which will allow us
to attempt pinning `Var`s on the first use, and force them to default types
otherwise.

This diff introduces a `Pin` binding, which will be used alongside
`Key::PinnedDefinition` for the cases where placeholder type pinning applies.

It is unused at this stage.

Reviewed By: yangdanny97

Differential Revision: D76273080

fbshipit-source-id: 56845767a4d212ffbebcda0a6946d65a92c3fbbb