Add a test demonstrating `Var` leakage through chained first use
Summary:
At this point I believe that I have almost eliminated nondeterminism originating
from the use of placeholder type `Var`s: at this point only a `NameAssign` is
allowed to create unpinned placeholders, and only the first use can "see" them
so the mutation is reasonably controlled.

There remains one known gap: because pinning is all-or-nothing (at this point
we have no notion of the "origin" of a `Var` so there's not an obvious way
to pin selectively), the fact that we *don't* pin on a `NameAssign` means that
we can leak `Vars` into chains of first-usages in ways that can still pin
nondeterministically.

Add a new pair of tests for this: here, `x` has a placeholder type and `w`
is the first use, which means the pinning doesn't happen when `w` evaluates.
As a result the *second* use of `x` will get `Any` if it evaluates early,
because it will use `Pin(x, ...)` which will pin the type to `Any`. But if
`z` (which is the first use of `w`) evaluates first, it will still see the
`Var` and we'll instead pin to `str`.

My plan is to fix this with one more layer of indirection - only needed
for first-uses that are also NameAssigns - to ensure that the `Pin(...)`
bindings evaluate in a known order.

Reviewed By: yangdanny97

Differential Revision: D76273090

fbshipit-source-id: 77cae6eddab856462831777a89139f12fa74f015