Implement minimal Pin logic for name assigns (regression!)
Summary:
NOTE: I have this marked as WIP because it breaks 2 lsp tests; most likely
there's somewhere that I need to treat `Pin` like a `Forward` and keep searching
up the dependency chain for a definition.

This diff implements the indirection to pin placeholder types in `NameAssign`,
by creating an extra `Key::PinnedDefinition` => `Binding::Pin` pair
that will handle taking the raw, unpinned result and pinning types (attempting
to do so by calculating the first use, and if that does not pin the `Var`s
then forcing them to default values).

At this stage of the implementation we have wired up the two bindings, but we
aren't actually handling first-usage lookups so there's no (intended) behavior change,
the `Binding::Pin` is basically just acting as a `Forward` right now.

There is actually a behavior change: a very simple loop test (which was originally just
an lsp hover test, but I've also turned it into a new test case in `delayed_inference.rs`)
changes. The regression is not great, but I suspect we should take it anyway for now,
I think it is pointing at loop recursion issues (see more discussion below).

There's also a small regression in autocomplete - for reasons I can't work out, we are now
getting two copies of a local variable. I suspect something in the available definitions logic
isn't working the way I expected; it should be the case that the scope sees exactly one
of the `Key::Definition` or `Key::PinnedDefinition`, which I tried to handle in D76273082,
but I think it must be finding both in at least one case. I think we should probably accept
this regression and investigate at the top of the stack, it's a pretty big distraction from
determinism work for me to be wading into this (thoughts kinto0, SamChou19815?) and
my understanding is grievejia is considering changing how these lookups work anyway.

---------

**Explaining the regression Regression**

We have a pretty nasty regression in loop handling: we now get an `Unknown` in this test:
```
x = 5
while True:
  x = x + 1
y = x
assert_type(y, Literal[1] | int)  # E: assert_type(Literal[5] | Unknown, Literal[1] | int)
```

I have a partial answer for why it happens: the type of `x` in the loop is part of a cycle, and
the result is not entrypoint invariant:
- if we hit recursion on the `Phi`, we'll get a `Var::Recursive` with a default of `Literal[1]`
- if we hit recursion on either of the two bindings for `x = x + 1`, we'll get no default

From a lot of print debugging, I can see that we are somehow encountering the `Pin`
version of the `x = x + 1` binding first. This causes us to first run around the cyle (through
the unpinned raw binding and then the Phi) and then get a recursive var with no
default.

It then gets forced to `Unknown` (which is not what would happen for a recursive var
introduced by the `Phi`), which propagates through the cycle.

I'm fuzzy on exactly why the `Pin` is getting computed first, intuitively I would have expected
the `Phi` to be, but regardless this is pretty clearly pointing at issues with loop recursion and
lack of cycle invariance. Given that we know pinning placeholder `Var`s (which are all the `Var`s
*except* the recursive ones) doesn't fix nondeterminism on some large projects we care about,
my guess is that we're going to have to dig further, and my current opinion is that we should
accept this regression (and maybe even be glad we have a very simple test case to investigate further!)

Reviewed By: yangdanny97

Differential Revision: D76273083

fbshipit-source-id: 44c74c656d5beb278ecdfe265d47e10eb6d03835