Add `Key::PinnedDefinition`
Summary:
This diff kicks off implementing pinning of placeholder types for
name assignments by introducing a key that will serve as an identity
for the pinned version of a definition.

The idea is that, in cases where the definition might introduce placeholder
types, we will not put the raw definition in any scopes - only the first usage
will ever be able to see the raw definition. All *other* usages will flow
through a pinned definition that depends on the first usage (if there is one),
and forces any placeholder types that are not pinned by the end of that first
usage.

Reviewed By: yangdanny97

Differential Revision: D76273082

fbshipit-source-id: 178fe2d587e2558fab48e1c70e516ff050f074ce