Make `User` contain a `Usage`
Summary:
So close the remaining gap in var determinism (chains of first-usage
name assigns), I need to start collecting reverse dependencies from
lookups, so that I can create an extra binding that depends on them
(which will avoid the chaining where unpinned Vars can be forwarded
along mutiple first usages and be pinned nondeterministically).

To do that, my plan is to make `Usage::User` able to track them, and
have the code pass a `Usage` by ref so that we can add the reverse
dependencies. That requires the owned value to live as long as the
binding we are creating - the obvious way to do it is to put that data
in the `User`.

Before I can change to passing by ref, I first need to change how `User`
is represented; this diff makes the change: it now wraps a `User::Usage`.

Reviewed By: yangdanny97

Differential Revision: D76273093

fbshipit-source-id: 9d0d6a6415301d8fa67feb842e6eaa196047e642