PEP695 variance inference (currently just for base classes)
Summary:
Variance inference algorithm which only processes base classes for now.
The algorithm itself closely follows what I did in pyre1 https://www.internalfb.com/code/fbsource/[b7c0c3769039]/fbcode/tools/pyre/source/analysis/attributeResolution.ml?lines=72

Things to do next:
- thread field info (discussed with Steven offline. This will take a non trivial bit of pluming because just getting the raw type on the fields isn't accurate)
- finish the rest of on_type

Two things that I am hoping for more comments on:

**- class type vs parametric type**
 One major difference is that in pyre1, we had a "parametric type" vs in pyrefly, we have a "class type". class type seems to be a superset of parametric types. so in examples like this one

`class Counter[T](dict[T, int]): ...`

we wind up trying to lookup things like `int` in the variance map, which don't end up being found.  I think if we only look into class types with non-empty type params, that should be equivalent to a parametric type?

**- Class heirchy:**

Consider this program:

    class A[T]:
        x: T

    class B[T](A[list[T]]):
       y: T

we need to be able to find the object associated with this list when we call `on_edge`.

The list isn't a child of any class in this program, so we can't just look at the ancestors.

In pyre1, we had a way to lookup a class via its name for the entire hierchy of the program.

**Workaround:**
Wonder if I need carry a map from that goes from string to classes encountered so far? this way for the example above, I am storing a map containing "list -> the list object".

Intuitively, for inference, we need 1- all base class info and 2- the class info encountered in the program

so it could be that all we need is the ability to look the ancestors of a class + all the classes we encountered in the program so far which can be stored in a map

Reviewed By: yangdanny97

Differential Revision: D75761324

fbshipit-source-id: f017970257a0daf85ff1a175165fe3e80e0d3523