Use Pyrefly config finding default logic before langauge client provides expected python environment (#451)
Summary:
Pull Request resolved: https://github.com/facebook/pyrefly/pull/451

vscode-python is only used in VSCode, so our logic around getting the interpreter from the IDE only really works there. This was reported as part of the problem in [this issue](https://github.com/facebook/pyrefly/issues/356), where Helix is used but Pyrefly isn't finding a local .venv interpreter.

To fix that, we should just not set a default interpreter in the IDE until one is provided to us. We should instead use the directory the `standard_config_finder` gives us and try to find an interpreter at that location.

If a language client later gives us an interpreter to use, this will be fine since we should clear out any cached configs and start using the newly provided interpreter in the workspace it was configured in.

Reviewed By: grievejia

Differential Revision: D76306932

fbshipit-source-id: 0041f679d371f9087d24e2a5dbe84609da3207eb