Add pre-commit hook and documentation
Summary:
Add a pre-commit hook and documentation for developers to use it

Fixes: https://github.com/facebook/pyrefly/issues/224

This solution is not as ideal as having a full standalone pre-commit package that is installable like this: https://github.com/astral-sh/ruff-pre-commit

We assume that you have pyrefly installed already and available on the path. This might actually be okay, since most people will want to write their own configs anyways and pin to a specific version.

EDIT: Talking to more people, `system` install might actually be the ideal situation, since a formatter would not need dependencies to run. Typechecking needs depedency and config.

Reviewed By: yangdanny97

Differential Revision: D76135389

fbshipit-source-id: 1e7be29bb911fe2e49db2bc6fe501bfc551ae455