Re-parse destination module for kwarg gotodef if the AST is not available
Summary:
As pointed out in D76379065: we relied on the availability of an AST at destination module for location refinement, but the AST may not be there if the destination module is not opened.

This diff makes it such that if the AST is not already available, grab the string content of the destination module and re-parse it on the fly. This way, cross-module kwarg gotodef location refinement could more much more reliably.

Reviewed By: SamChou19815

Differential Revision: D76448736

fbshipit-source-id: c7f1f615900247d175a544d7ac8da6cda0042494