Refine keyword argument gotodef location
Summary:
In D76379066, we handle kwarg gotodef by jumping to the function name of the callee. That is roughly where the user wants to go, but it would be nice if we could further pinpoint the exact location where the formal param of the same name is located.

This diff "refines" the callee location, by performing another AST walk (i.e. issuing a covering node request) to locate the corresponding `StmtFunctionDef` AST node for the callee. After that, we inspect the syntax of the function and filter out names based on the name of the argument.

Reviewed By: SamChou19815

Differential Revision: D76379065

fbshipit-source-id: 46807354979e6dbe3861c58a96e66606024a0365