call simplify_mut in expand_mut
Summary:
This  diffmakes expand_mut simplify after substituting vars with their bound values, similar to deep_force_mut.

I think this is a better solution to D76443381, since we already call expand on the return type when we evaluate a call.

There was another call site that does some ad-hoc simplification of unions, which is no longer necessary.

Reviewed By: ndmitchell

Differential Revision: D76454633

fbshipit-source-id: 8fd09047a3d56ef26d1f5a33137ac910047b27b6