Look through `AssignToAttribute` and `AssignToSubscript` bindings when resolving definitions
Summary:
I noticed that if we have multiple assignments to an attribute or a subscript, we were not able to correctly resolve gotodefs on the base identifier after the first assignment. E.g.
```
def test(x: list[int]) -> None:
    x[0] = 1
#   ^ This works
    x[1] = 2
#   ^ This doesn't work
    x[2] = 3
#   ^ This doesn't work
```
The issue here seems to be that after the first assignment, every use of `x` going forward always point to a narrowing binding `AssignToSubscript` (and `AssignToAttribute`, resp.), and our gotodef logic isn't taught how to interpret those bindings and assumes they can't point to a definition -- but they actually can if we correctly take apart the binding and find the right identifier in it to construct a forwarding use from it.

This diff copies over from the solving logic how to correctly interpret these special narrowing bindings, to re-establish the broken link towards the actual definition.

Reviewed By: SamChou19815

Differential Revision: D76470485

fbshipit-source-id: 719ed20c8ec1c2832b0f7666541a9dc5f6874c32