Increase initial gas size
Summary:
gasworks

I've observed several occasions in offline tests in IG where gotodef would succeed if I remove a few line of code, but would fail if I don't. Furthermore, I don't have to remove the same lines of code -- as long the removed code is relevant to the gotodef variable, and size of the code shrinks, gotodef would succeed.

I think this strongly suggests that the issue has something to do with our initial gas size too small. The diff proposes we increase the max number of hops from 20 to 100.

Reviewed By: kinto0

Differential Revision: D76467435

fbshipit-source-id: 8af366acdc2e5381f7cb61d9f57d4ea21690c080