Demonstrate nondeterminism from loops
Summary:
For a while now I've been working to implement first-use tracking for the pinning of
"placeholder types", by which I technically mean any `Type::Var` for which the `Variable`
is not either an `Answer` or `Recursive`.

I believe I have largely finished this in a stack, but we are still seeing quite a lot
of nondeterminism in some example projects. I strongly suspect that this comes from
cycles not being well-behaved: exports can cause us to enter a cycle at different
places, and there's no a priori reason to think the results will always come out the
same.

This example proves that in at least some cases, recursion does indeed
introduce nondeterminism: I can import `y` or `z` to enter the cycle defining `x`
at different points, and get different inferred types. The example is similar to
some of the placeholder-type-based examples, although I personally find it quite
a bit harder to actually predict what will happen when looking at loop recursion
examples (this isn't so surprising: if there is any nondeterminism, it will
propagate out not from the binding where we entered the cycle, but from the last
binding *before* it, since that's the first one to see a recursive answer. That's
pretty tricky to reason about!)

Reviewed By: yangdanny97, ndmitchell

Differential Revision: D76304255

fbshipit-source-id: 47b46e9ceefcbbd6d55449ab5aa376f2527d4e1d