Do not crash on failed definition key lookup
Summary:
We have this quirk in Pyrefly where if the code is protected by `if TYPE_CHECKING:` or `if sys.version...`, then we'd hide the corresponding AST node from the binding phase. This means that when IDE tries to look up a def or use under those hidden AST node, the lookup is going to fail and our internal logic is to crash immediately when that happens.

There's pre-existing guard for use keys against this kind of lookup failures. But we haven't done it for def keys, which I ended up observing relevant crash in my offline experiments. Let's fix that.

Reviewed By: kinto0

Differential Revision: D76524697

fbshipit-source-id: 0c7313f7fbd3a48cab0700af06bc3fde18dc2d39