Add overload case to variance inference
Summary:
We should consider overloads in our variance inference by interating over all the signatures and applying `on_type` to each signature.

This is also how we handled it in pyre1, except that overloads were not a type.

One unit test result changed to the correct result (see D75761324 for context on that unittest).

Reviewed By: yangdanny97

Differential Revision: D76452375

fbshipit-source-id: 580dfa8ac8e606cbadb695b9920f974b8953b88a