Add readonly semantics into the algorithm
Summary:
For readwrite, we start as invariant, while for readonly (and methods) we start as covariant.

We don't have a great way to distinguish between fields and methods right now, so we should fix that soon.

Steven and I chatted about this and we agreed that it is ok to consider non descriptor callables as methods for now, but there should be a better solution to this.

Additionally, those fields are currently marked as read-write, which is why we need the distinction.

Reviewed By: rchen152

Differential Revision: D76477029

fbshipit-source-id: e10a76876ec1987d23368a19cfe46de91f350e1e