Cleanup variance binding by removing unnessesairy fields and base classes
Summary:
We know that fields and base classes for the class we start variance algorithm with, is not enough.

Instead, we needed to have a mechanism that grabs fields and base classes to any class, given the `class`.

Since already have this mechanism (through constructing the class field key for fields, and through accessing the bases of a class from the metaclass data), we don't need those two things in the binding anymore.

Reviewed By: rchen152

Differential Revision: D76536414

fbshipit-source-id: b35629f2d43b76bd95c3e78d5637af9b9c160d10