Add a test for contextual typing against unions (works!)
Summary:
We were discussing this as one of the features we want soon, and we
were having trouble remembering whether it is already implemented (looks
like it is).

I couldn't find a well-marked unit test, and I actually confused myself
the first time I tried it (forgot the parentheses in `B()`, which led
to a type error and made me think it doesn't work) so I think it's
good to have a test tracking this.

Reviewed By: grievejia

Differential Revision: D76558293

fbshipit-source-id: 95001d1e96da068da522f008e59c789daa95d5e4