Extend union contextual typing test to flip union order (still works)
Summary:
Whether union ordering affects contextual typing came up in discussion.

It appears to work fine at least for a simple union of size 2 in either order;
I also added cases where there are lists on both sides of the union to make sure
that we aren't special-casing a structural mismatch.

As far as I can tell from any of these examples it works as we we would like.

Reviewed By: rchen152

Differential Revision: D76559685

fbshipit-source-id: 0b39241125f69063ebf79e039fb9931ce250abb6