Add a test case with poor first-use behavior
Summary:
This test case shows that Pyrefly behaves badly right now if the first use
reads the same name twice.

The root cause of the problem is that the first read of the name uses the
raw idx, but the second one uses the pinned idx. This causes problems
because the pinned idx depends on the first use, so we cause a cycle
(which could lead to perf issues, but it also has bad type checking behavior
- we might try to pin the recursive `Var` created by the cycle inconsistently).

Reviewed By: ndmitchell

Differential Revision: D76693210

fbshipit-source-id: 93f24aab213ae552caaa76c173121f296cc55cfd