Fix bug related to multiple reads from first use
Summary:
We need to make sure that the first use binding uses the raw binding with
placeholder types throughout (so if we look up the name multiple times, keep
using the raw idx even though this isn't the first hit on a lookup).

We can do that by comparing the current `usage` to whatever was recorded
in the `Pin`.

Reviewed By: ndmitchell

Differential Revision: D76693316

fbshipit-source-id: bd4ea9060b55ad0d4682a872b37a2989596d2c09