Make it clearer when cycle breaking is happening
Summary:
The use of `AnnAssign` as the type check kind (and an error message that is optimized
for `AnnAssign` and *intentionally* tells us nothing whatsoever about where the type
came from!) made it pretty hard to debug a scenario where Pyrefly was inappropriately
using cycle breaking and then throwing an error when the cyclic calculation finished.

Let's use a dedicated error kind, and include some kind of indication in the error
message (this error message still won't be great for users, but it sure beats no
hint at all, and more importantly it could save us a lot of time both debugging
Pyrefly bugs that *cause* cycles as well as answering user questions when this happens
in production).

Reviewed By: ndmitchell

Differential Revision: D76693286

fbshipit-source-id: b4a94cce6e62e10966695f1f6eb218cf356dabc2