Demonstrate cyclic decorators
Summary:
I've been giving thought to whether type annotations are sufficient
to eliminate nondeterminism in `Binding` cycles. I think they might be,
for valid code.

But I was reminded that Pyre1 was forced to put a cycle detector into
our decorator logic (it's part of the giant `AttributeResolution` tangle),
and realized it's pretty easy to write code in an import cycle where
we hit this problem; our nondeterministic cycle breaking then causes
behavior differences over data races because whichever export is
requested first will be treated as `Any` by its direct dependent in
the cycle.

I'm not sure how important this is - the code exhibiting the problem
can certainly appear in stubs or in code that's in the process of being
written, but at runtime it would produce an actual cycle and crash.

Nontheless, it's good to be aware of, particularly if/when we're ready
to attempt to eliminate nondeterminism.

Reviewed By: yangdanny97

Differential Revision: D76602787

fbshipit-source-id: f20ca46762745eafe5cf2dffed0c3d3584ffacc6