Remove `NamedBinding::binding`
Summary:
This field is currently not used anywhere (it's explicitly marked as "dead code"). Although it's not impossible to properly preserve it in D76755128, that's additional work and I don't see a strong reason why preserving the field is important.

I feel that we might want to prioritize fixing semantic tokens with D76755128 (since we are about to release it to users very soon) and if we really want the binding, I'd happy to stack on top to bring it back.

Reviewed By: SamChou19815

Differential Revision: D76755126

fbshipit-source-id: f61d5353725d42d706ee0ec1957576cbaeefef84