Remove Clone from AnswersSolver
Summary:
I don't think there's ever a good reason to clone one of these, I
don't think we should be deriving Clone even if we can.

I want to be able to think about it as one answers solver per
entry into a module per thread, which is the case but having an
unused Clone derivation makes that less obvious.

Reviewed By: grievejia

Differential Revision: D76762194

fbshipit-source-id: 3c68dcf4b7d620149114cb2679469f98edcc0bd3