Sometimes exit immediately on a config parse error in per-file mode
Summary:
Fixes https://github.com/facebook/pyrefly/issues/457.

We can't always exit immediately on an unparseable config in per-file mode, as we discover configs on the way as we type-check. But when a config is explicitly passed in via `-c` or we're type-checking only a single file/directory, we can.

Reviewed By: stroxler

Differential Revision: D76779356

fbshipit-source-id: cf72ae5a97a02d0de08c6bac7b9225f1df0a34d2