Create CalculationStack to enable better cycle / idx debugging
Summary:
This diff begins work on making it easier to see what is happening
in cycles by threading a `CalculationStack` through `AnswersSolver`.

This diff by itself is sufficient to get at the current `Idx` in
debugging code or to debug a cycle directly in `main` in a unit test,
but it is incomplete and not ready for debugging large production
codebases with cycles because the stack is per `AnswersSolver`, and
we switch solvers every time we change modules.

Reviewed By: ndmitchell

Differential Revision: D76762188

fbshipit-source-id: ce1f24e9f017f65e82db28b86f7c3707d942d7ca