Add an AnyIdx type to enable general handling of graph node identities
Summary:
I want to add the ability for `AnswersSolver` to track what is currently
being worked on by the current thread. I think this will help debugging
quite a bit in some cases.

The most common scenario is where we just want to know the key *currently*
being solved - I want to set a `debug` flag based on this at least twice a week
and it's unavailable. But knowing *all* of the keys currently in progress
would be even better, particularly for debugging cycle handling (relevant to
determinism efforts on example codebases today, also relevant for anyone
trying to change loop variables or dealing with bugs in `Pin` bindings).

In order to track the stack of things we're working on we need to either work
with `& dyn` or have a single, concrete static datatype. I think the concrete
datatype is preferable, and that's what `AnyIdx` gives us.

Reviewed By: grievejia, ndmitchell

Differential Revision: D76762190

fbshipit-source-id: baceedbf292d5eab995f8d5880f877148d8f0602