Add the ability to see the current idx or binding in debugging
Summary:
It is pretty easy to get down into recursive solving of an expression or
a complex binding and be somewhere in the call stack where we don't have
direct access to the binding anymore, so `show_current_binding` can be handy.

And prior to the implementation of `CalculationStack` it was actually not
possible to figure out the current idx being solved at all once we entered
the solve, because that information disappeared into the stack frame of
`get_idx`.

This diff makes both of these possible to easily see with debugging tools,
and along the way it implements `show_binding_for(idx)` which is something that
I've wanted several times... for example just yesterday I was debugging a
a `Pin` binding and had to chain things together to figure out what the
pin was actually pointing at; a one-stop helper makes it quicker which is
important for debugging velocity.

Example of using this:
```
> hg diff
 diff --git a/pyrefly/pyrefly/lib/alt/expr.rs b/fbcode/pyrefly/pyrefly/lib/alt/expr.rs
 --- a/pyrefly/pyrefly/lib/alt/expr.rs
+++ b/pyrefly/pyrefly/lib/alt/expr.rs
@@ -1107,6 +1107,14 @@
                 }
             }
             Expr::List(x) => {
+                let debug = self.show_current_module() == "main";
+                if debug {
+                    println!(
+                        "Inside list literal, current idx / binding are {} / {}",
+                        self.show_current_idx(),
+                        self.show_current_binding()
+                    )
+                }
                 let elt_hint = hint.and_then(|ty| self.decompose_list(ty));
                 if x.is_empty() {
```
and then
```
cargo test test_side_effecting_comprehension_targets -- --nocapture
```
(I just picked some test where I thought a list would be involved in some
binding where it wouldn't be easy to see the current binding without the new
debug tools) which gave me:
```
Inside list literal, current idx / binding are main . Key :: anon 900..909 / iter [1, 2, 3]
Inside list literal, current idx / binding are main . Key :: anon 930..939 / iter [1, 2, 3]
Inside list literal, current idx / binding are main . Key :: anon 1023..1038 / iter ["1", "2", "3"]
Inside list literal, current idx / binding are main . Key :: anon 1122..1137 / iter ["1", "2", "3"]
ERROR `str` is not assignable to attribute `d` with type `int` [bad-assignment]
   --> pyrefly/lib/test/assign.rs:669:12
    |
669 |     [1 for c.d in ["1", "2", "3"]]  # E: `str` is not assignable to attribute `d` with type `int`
    |            ^^^
    |
ERROR Cannot set item in `C` [bad-argument-type]
   --> pyrefly/lib/test/assign.rs:670:12
    |
670 |     [1 for c[0] in ["1", "2", "3"]]  # E: Argument `str` is not assignable to parameter `value` with type `int` in function `C.__se...
    |            ^^^^
    |
  Argument `str` is not assignable to parameter `value` with type `int` in function `C.__setitem__`
test test::assign::test_side_effecting_comprehension_targets ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 1570 filtered out; finished in 1.25s

     Running unittests bin/main.rs (/data/users/stroxler/fbsource/fbcode/pyrefly/target/debug/deps/pyrefly-038944b2672dd122)
```

This is pretty nice, it's great to be able to see the currently in progress
key / binding pair when debugging!

I actually was sort of expecting it to be computing the binding for the overall expr - I had
forgotten that `Binding::Iter` is its own thing - so this actually slightly surprised me... which
is exactly why it's so useful for this to be easy to do :)

Reviewed By: ndmitchell

Differential Revision: D76762189

fbshipit-source-id: 445f5e0a88eb56e6218dca11ce48b8376cfea1bd