Port all pyre variance tests to pyrefly
Summary:
To ensure completeness, I'm porting all unittests from pyre1 to pyrefly, including a couple of cases that are not in the conformance suite.

I will followup with some todo of things I'd like to follow up on, but they won't be the main priority, so I will mostly update them as I go.

Reviewed By: ndmitchell

Differential Revision: D76777048

fbshipit-source-id: 9a1a1285f42563f4b95282869e448753a8c8741d