improve invalid type annotation errors
Summary:
"Expected a type argument for <something that can't be subscripted>" is confusing. I add a check to see which special forms can actually be subscripted, and emit a different error for the ones that can't.

I ended up also falling into a rabbit hole of adding backticks in a lot of error messages. Probably should have split it up but it's too difficult to untangle now.

Reviewed By: ndmitchell

Differential Revision: D76870823

fbshipit-source-id: ac7b357462e429c56339dd6529ee71bb2d5a415a