Ensure you bind loop variables properly
Summary: The first trip around the loop (without breaks) must be bound into the right phi keys. Previously we got it wrong, but mostly it didn't matter because there were suitable phi keys in lots of other places. If you engineer a case where there is termination in the `orelse` that doesn't happen. See the attached test cases.

Reviewed By: stroxler

Differential Revision: D76892883

fbshipit-source-id: db2da70e13f7a4e798441be5813a0155f6c4927b