Make sure we ensure the RHS of invalid assignments
Summary:
Fixes https://github.com/facebook/pyrefly/issues/518

If we had an invalid identifier, we forgot to binding the RHS. For walrus assignment, that was a problem, as we did evaluate the RHS. For normal bindings it didn't crash, but we did forget to check them, which sucked. Now check them properly.

Reviewed By: stroxler

Differential Revision: D76892882

fbshipit-source-id: 051f9fc3d92c0f2de76d1ea20487aace46474204