Add failing test about TypedDict not supporting pop()
Summary:
`pop` is not working for `TypedDict` because there is no synthesized method for it. Will look into synthesizing the method since there's a TODO about it in the code.

Haven't looked at this before so still piecing everything together.

Initial thought:
It seems that we already have a typeshed signature for pop, https://github.com/python/typeshed/blob/main/stdlib/typing.pyi#L1013C1-L1014C1 (as well as __delitem__, __or__ and other things mentioned in the TODO)

So unclear to me yet why we have to have a typeshed signature + a synthesized method.

Will look into that next.

Reviewed By: rchen152, ndmitchell

Differential Revision: D76842159

fbshipit-source-id: 92fb91b660f525aceda286be0c4ef890ed24c8c9