Remove `bug` marker on test
Summary:
Noticed this while implementing cycle breaking.

FWIW the super code is not working well - it seems to be very sensitive
to cycles, and in fact a bunch of cases that flipped when I implemented var
pinning wind up flipping back again.

I suspect the root cause is that the logic needs to be more aware of
cycle breaking (but obviously until we break them predicatably, trying to
accomplish this is very difficult!)

Reviewed By: rchen152

Differential Revision: D76944309

fbshipit-source-id: bd25ea6a47cc019c54a55bbaa19f106d6bdc5d1c