Limit the number of facets we store
Summary:
It used to be unbounded. For files like z3.z3core we generated fairly huge facets, and they took O(n^2) perf, as each line took more and more time to clone. This change reduces from 1.2s to 0.3s.

If we drop the limit to 10 it reduces to 0.08s, so possible that 100 is still on the large size. But let's start there.

Reviewed By: stroxler

Differential Revision: D76919353

fbshipit-source-id: e43d8c3e2c781d8a6632f15fbb8e9a358abb3b84