Ensure init works if you use self in init
Summary:
Fixes https://github.com/facebook/pyrefly/issues/444

When you use `self` in `__init__` you end up creating a cycle. That's fine, the cycle resolves perfectly reasonably to the right result. But when we were creating bound methods we match on the type, and if it was a Var we didn't expand it. Fix that.

Reviewed By: yangdanny97

Differential Revision: D77018929

fbshipit-source-id: e00d10c56d391a0f36d40fbae17f2057c7a9b470