Parallelize export searching for auto import
Summary:
Currently, autoimport will scan through the export table of the entire project. While this is relatively cheap, the costs do add up.

This diff optimizes it by parallelizing the work. Notably, I will the work into unit of 1k modules, so that we don't have to pay the locking cost for each single module.

Reviewed By: ndmitchell

Differential Revision: D76924171

fbshipit-source-id: 3085c6458b91a122de604fd7b7982234aef3b252