filter out INFO diagnostics when counting errors & determining exit code (#525)
Summary:
I think this is simpler - no need to deal with a separate error collector.

We exclude reveal-type diagnostics from the displayed error count & when determining exit codes. We also do not add an error suppression comment for any reveal-types. Warnings are still treated the same as errors, for now.

Closes https://github.com/facebook/pyrefly/issues/506

Pull Request resolved: https://github.com/facebook/pyrefly/pull/525

Reviewed By: ndmitchell

Differential Revision: D77032158

fbshipit-source-id: 778645a762d8771d873de847131f02364101032d