Disable test_completion temporarily
Summary:
I have a diff (D77031603) blocked on this test misbehaving, I'm not
sure if the result is inherently nondeterministic or is just system-dependent,
but I have local and CI results diverging.

I also just noticed that external CI is failing with the same mismatch, so this
is almost certainly something going wrong in the lsp test itself rather than
related to the diff being blocked.

Reviewed By: SamChou19815

Differential Revision: D77034910

fbshipit-source-id: d9b6fd3d9a3e9c8b2b82cd5bc1287e1ffbdb3e59